{{cookiecutter.repo_name}}/requirements.txt
Django>=1.9,<1.10
waitress>=0.8,<0.9
whitenoise>=2.0,<2.1

dj-database-url==0.3.0
django-debug-toolbar>=1.4,<1.5
django-sslify==0.2.7

python-decouple==3.0
pytz==2015.7
sqlparse>=0.1,<0.2

{% if cookiecutter.use_postgres == 'y' -%}
psycopg2>=2.6,<2.7
{%- endif %}

{% if cookiecutter.use_redis == 'y' -%}
redis>=2.10,<2.11
django-redis>=4.3,<4.4
{%- endif %}

{% if cookiecutter.use_rq == "y" -%}
rq>=0.5,<0.6
django-rq>=0.9,<0.10
{%- endif %}

{% if cookiecutter.use_sentry == "y" -%}
raven>=5.16,<5.17
{%- endif %}
